feat: 스케쥴 통합과 쓰기 요청 정리

+++ ModbusClient/src/main/java/com/example/ModbusClient/config/netty/NettyClient.java
@@ -3,6 +3,7 @@ package com.example.ModbusClient.config.netty;
 import com.example.ModbusClient.config.ServerConfig;
 import com.example.ModbusClient.entity.modbus.ServerInfo;
 import com.example.ModbusClient.service.ModbusServiceTest;
+import com.example.ModbusClient.service.MqttConvertTCP;
 import io.netty.bootstrap.Bootstrap;
 import io.netty.channel.Channel;
 import io.netty.channel.ChannelFuture;
@@ -10,6 +11,8 @@ import io.netty.channel.ChannelFutureListener;
 import jakarta.annotation.PreDestroy;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
@@ -27,6 +30,7 @@ public class NettyClient {
     private final ServerConfig serverConfig;
     private final ModbusServiceTest modbusService;
 
+
     public void start() {
         List<ServerInfo> servers = serverConfig.getServerList();
         for (ServerInfo serverInfo : servers) {
@@ -43,11 +47,15 @@ public class NettyClient {
                     log.info("Connected to server: {}:{}", host, port);
                 }
                 modbusService.addServer(channel);
+                modbusService.startScheduling();
             } else {
                 log.error("Failed to server: {}:{}", host, port);
                 channel.closeFuture().addListener((ChannelFutureListener) future1 -> {
                     log.warn("Connection closed. Attempting to reconnect...");
                     attemptReconnect();
+                    if (future1.isSuccess()) {
+//                        modbusService.heartbeatRequest();
+                    }
                 });
             }
         });
@@ -70,9 +78,17 @@ public class NettyClient {
     @PreDestroy
     public void stop() {
         if (channel != null) {
-            channel.close();
-            channel.parent().closeFuture();
-            modbusService.removeServer(channel);
+            channel.close().addListener((ChannelFutureListener) future -> {
+                if (future.isSuccess()) {
+                    modbusService.removeServer(channel);
+                }
+                channel.parent().close().addListener((ChannelFutureListener) parent -> {
+                    if (parent.isSuccess()) {
+                        log.info("Connection closed. Attempting to reconnect .....");
+                        attemptReconnect();
+                    }
+                });
+            });
         }
     }
 }
